--- a/react/client/base.jsx
+++ b/react/client/base.jsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 export function createApp () {
   const wrapperWidth = 960
   const wrapperHeight = 720