--- a/react/package.json
+++ b/react/package.json
@@ -12,8 +12,8 @@
   "dependencies": {
     "@fastify/vite": "^6.0.7",
     "fastify": "^4.28.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "react": "19.0.0-rc-f90a6bcc-20240827",
+    "react-dom": "19.0.0-rc-f90a6bcc-20240827",
     "ws": "^8.18.0"
   },
   "devDependencies": {