## Stacked on https://github.com/platformatic/ssr-performance-showdown/pull/7

This is maybe not fair to merge since it's a prerelease but just wanted to put it out there since it's a mild improvement.

This upgrades to 19 and switches to modern JSX transform.

## Before (assuming https://github.com/platformatic/ssr-performance-showdown/pull/7 lands)

```
┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬───────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max   │
├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼───────┤
│ Latency │ 11 ms │ 12 ms │ 30 ms │ 36 ms │ 13.81 ms │ 5.61 ms │ 56 ms │
└─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴───────┘
┌───────────┬────────┬────────┬────────┬────────┬────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg    │ Stdev   │ Min    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Req/Sec   │ 658    │ 658    │ 698    │ 719    │ 700.6  │ 16.99   │ 658    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Bytes/Sec │ 129 MB │ 129 MB │ 137 MB │ 141 MB │ 138 MB │ 3.33 MB │ 129 MB │
└───────────┴────────┴────────┴────────┴────────┴────────┴─────────┴────────┘
```

## After

```
┌─────────┬───────┬───────┬───────┬───────┬──────────┬─────────┬───────┐
│ Stat    │ 2.5%  │ 50%   │ 97.5% │ 99%   │ Avg      │ Stdev   │ Max   │
├─────────┼───────┼───────┼───────┼───────┼──────────┼─────────┼───────┤
│ Latency │ 10 ms │ 11 ms │ 30 ms │ 36 ms │ 12.67 ms │ 5.45 ms │ 55 ms │
└─────────┴───────┴───────┴───────┴───────┴──────────┴─────────┴───────┘
┌───────────┬────────┬────────┬────────┬────────┬────────┬─────────┬────────┐
│ Stat      │ 1%     │ 2.5%   │ 50%    │ 97.5%  │ Avg    │ Stdev   │ Min    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Req/Sec   │ 712    │ 712    │ 769    │ 780    │ 762.8  │ 19.06   │ 712    │
├───────────┼────────┼────────┼────────┼────────┼────────┼─────────┼────────┤
│ Bytes/Sec │ 140 MB │ 140 MB │ 151 MB │ 153 MB │ 150 MB │ 3.74 MB │ 140 MB │
└───────────┴────────┴────────┴────────┴────────┴────────┴─────────┴────────┘
```