@slorber 2024-08-28
For those interested in knowing why it improves: see the [React "Fast JSX" PR](https://github.com/facebook/react/pull/28768) here